refactor: example: envoy gateway
